+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 **go version**
 
 - feat: add `WithSwipeOffset` to set offset for swipe start/end point
+- feat: set random offset for tap/swipe points with `WithOffsetRandomRange`
 - change: set `WithOffset` deprecated, replace with `WithTapOffset`
 
 ## v4.3.6 (2023-09-07)